+++ api/v1alpha1/ingressgateway_types.go
@@ -29,6 +29,7 @@ func init() {
 
 // IngressGateway is the Schema for the ingressgateways API
 // +kubebuilder:printcolumn:name="Synced",type="string",JSONPath=".status.conditions[?(@.type==\"Synced\")].status",description="The sync status of the resource with Consul"
+// +kubebuilder:printcolumn:name="Last Synced",type="date",JSONPath=".status.lastSyncedTime",description="The last successful synced time of the resource with Consul"
 // +kubebuilder:printcolumn:name="Age",type="date",JSONPath=".metadata.creationTimestamp",description="The age of the resource"
 type IngressGateway struct {
 	metav1.TypeMeta   `json:",inline"`
@@ -167,6 +168,10 @@ func (in *IngressGateway) SetSyncedCondition(status corev1.ConditionStatus, reas
 			Message:            message,
 		},
 	}
+}
+
+func (in *IngressGateway) SetLastSyncedTime(time metav1.Time) {
+	in.Status.LastSyncedTime = time
 }
 
 func (in *IngressGateway) SyncedCondition() (status corev1.ConditionStatus, reason, message string) {