+++ api/v1alpha1/proxydefaults_types.go
@@ -29,6 +29,7 @@ func init() {
 
 // ProxyDefaults is the Schema for the proxydefaults API
 // +kubebuilder:printcolumn:name="Synced",type="string",JSONPath=".status.conditions[?(@.type==\"Synced\")].status",description="The sync status of the resource with Consul"
+// +kubebuilder:printcolumn:name="Last Synced",type="date",JSONPath=".status.lastSyncedTime",description="The last successful synced time of the resource with Consul"
 // +kubebuilder:printcolumn:name="Age",type="date",JSONPath=".metadata.creationTimestamp",description="The age of the resource"
 type ProxyDefaults struct {
 	metav1.TypeMeta   `json:",inline"`
@@ -133,6 +134,10 @@ func (in *ProxyDefaults) SetSyncedCondition(status corev1.ConditionStatus, reaso
 			Message:            message,
 		},
 	}
+}
+
+func (in *ProxyDefaults) SetLastSyncedTime(time metav1.Time) {
+	in.Status.LastSyncedTime = time
 }
 
 func (in *ProxyDefaults) ToConsul(datacenter string) capi.ConfigEntry {